Move the preferences panel to its own file.
