+++ src/components/Sidebar.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import {Color} from '../Color';
-import {preferences, PreferencesManager} from './PreferencesManager';
+import {PreferencesManager} from './PreferencesManager';
 import {LANGUAGES} from '../constants';
 import {TurmoilModel} from '../models/TurmoilModel';
 import {PartyName} from '../turmoil/parties/PartyName';
@@ -11,6 +11,7 @@ import {GlobalParameterValue} from './GlobalParameterValue';
 import {MoonGlobalParameterValue} from './MoonGlobalParameterValue';
 import {GlobalParameter} from '../GlobalParameter';
 import {MoonModel} from '../models/MoonModel';
+import {PreferencesDialog} from './PreferencesDialog';
 
 export const Sidebar = Vue.component('sidebar', {
   props: {
@@ -58,6 +59,7 @@ export const Sidebar = Vue.component('sidebar', {
     'game-setup-detail': GameSetupDetail,
     'global-parameter-value': GlobalParameterValue,
     'moon-global-parameter-value': MoonGlobalParameterValue,
+    'preferences-dialog': PreferencesDialog,
   },
   mixins: [TranslateMixin],
   data: function() {
@@ -85,62 +87,6 @@ export const Sidebar = Vue.component('sidebar', {
     };
   },
   methods: {
-    setPreferencesCSS: function(
-      val: boolean | undefined,
-      cssClassSuffix: string,
-    ): void {
-      const target = document.getElementById('ts-preferences-target');
-      if (!target) return;
-      if (val) {
-        target.classList.add('preferences_' + cssClassSuffix);
-      } else {
-        target.classList.remove('preferences_' + cssClassSuffix);
-      }
-
-      if (!target.classList.contains('language-' + this.lang)) {
-        target.classList.add('language-' + this.lang);
-      }
-    },
-    updatePreferencesFromStorage: function(): Map<
-            string,
-            boolean | string
-            > {
-      for (const k of preferences) {
-        const val = PreferencesManager.load(k);
-        if (k === 'lang') {
-          PreferencesManager.preferencesValues.set(k, this.$data[k]);
-          this[k] = val || 'en';
-          PreferencesManager.preferencesValues.set(k, val || 'en');
-        } else {
-          const boolVal = val !== '' ? val === '1' : this.$data[k];
-          PreferencesManager.preferencesValues.set(k, val === '1');
-          this.$data[k] = boolVal;
-        }
-      }
-      return PreferencesManager.preferencesValues;
-    },
-    updatePreferences: function(_evt: any): void {
-      let strVal: string = '';
-      for (const k of preferences) {
-        const val = PreferencesManager.preferencesValues.get(k);
-        if (val !== this.$data[k]) {
-          if (k === 'lang') {
-            strVal = this.$data[k];
-          } else {
-            strVal = this.$data[k] ? '1' : '0';
-          }
-          PreferencesManager.save(k, strVal);
-          PreferencesManager.preferencesValues.set(k, this.$data[k]);
-          this.setPreferencesCSS(this.$data[k], k);
-        }
-      }
-    },
-    syncPreferences: function(): void {
-      for (const k of preferences) {
-        this.$data[k] = PreferencesManager.preferencesValues.get(k);
-        this.setPreferencesCSS(this.$data[k], k);
-      }
-    },
     getPlayerColorCubeClass: function(): string {
       return this.acting_player && (PreferencesManager.loadBoolean('hide_animated_sidebar') === false) ? 'preferences_player_inner active' : 'preferences_player_inner';
     },
@@ -170,11 +116,8 @@ export const Sidebar = Vue.component('sidebar', {
       }
     },
   },
-  mounted: function() {
-    this.updatePreferencesFromStorage();
-  },
   template: `
-<div :class="'sidebar_cont sidebar '+getSideBarClass()" :data="syncPreferences()">
+<div :class="'sidebar_cont sidebar '+getSideBarClass()">
   <div class="tm">
     <div class="gen-text">GEN</div>
     <div class="gen-marker">{{ getGenMarker() }}</div>
@@ -227,107 +170,18 @@ export const Sidebar = Vue.component('sidebar', {
       <div class="info_panel_actions">
         <button class="btn btn-lg btn-primary" v-on:click="ui.gamesetup_detail_open=false">Ok</button>
       </div>
-  </div>
-</div>
-
-<a href="/help" target="_blank">
-  <div class="preferences_item preferences_item--help">
-    <i class="preferences_icon preferences_icon--help" :title="$t('player aid')"></i>
-  </div>
-</a>
-
-<div class="preferences_item preferences_item--settings">
-  <i class="preferences_icon preferences_icon--settings" :class="{'preferences_item--is-active': ui.preferences_panel_open}" v-on:click="ui.preferences_panel_open = !ui.preferences_panel_open"></i>
-    <div class="preferences_panel" v-if="ui.preferences_panel_open">
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="hide_hand">
-          <i class="form-icon"></i> <span v-i18n>Hide cards in hand</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="hide_awards_and_milestones">
-          <i class="form-icon"></i> <span v-i18n>Hide awards and milestones</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="small_cards">
-          <i class="form-icon"></i> <span v-i18n>Smaller cards</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="magnify_cards">
-          <i class="form-icon"></i> <span v-i18n>Magnify cards on hover</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="hide_discount_on_cards">
-          <i class="form-icon"></i> <span v-i18n>Hide discount on cards</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="show_card_number">
-          <i class="form-icon"></i> <span v-i18n>Show card numbers (req. refresh)</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="remove_background">
-          <i class="form-icon"></i> <span v-i18n>Remove background image</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="show_alerts">
-          <i class="form-icon"></i> <span v-i18n>Show in-game alerts</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="enable_sounds">
-          <i class="form-icon"></i> <span v-i18n>Enable sounds</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="hide_animated_sidebar">
-          <i class="form-icon"></i> <span v-i18n>Hide sidebar notification</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="hide_tile_confirmation">
-          <i class="form-icon"></i> <span v-i18n>Hide tile confirmation</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item">
-        <label class="form-switch">
-          <input type="checkbox" v-on:change="updatePreferences" v-model="learner_mode">
-          <i class="form-icon"></i>
-          <span v-i18n>Learner Mode (req. refresh)</span>
-          <span class="tooltip tooltip-left" data-tooltip="Show information that can be helpful\n to players who are still learning the games">&#9432;</span>
-        </label>
-      </div>
-      <div class="preferences_panel_item form-group">
-        <label class="form-label"><span v-i18n>Language</span> (<a href="javascript:document.location.reload(true);" v-i18n>refresh page</a> <span v-i18n>to see changes</span>)</label>
-        <div class="preferences_panel_langs">
-          <label class="form-radio" v-for="language in langs">
-            <input name="lang" type="radio" v-on:change="updatePreferences" v-model="lang" :value="language.id">
-            <i class="form-icon"></i> {{ language.title }}
-          </label>
-        </div>
-      </div>
-
-
-      <div class="preferences_panel_actions">
-        <button class="btn btn-lg btn-primary" v-on:click="ui.preferences_panel_open=false">Ok</button>
-      </div>
     </div>
+  </div>
+
+  <a href="/help" target="_blank">
+    <div class="preferences_item preferences_item--help">
+      <i class="preferences_icon preferences_icon--help" :title="$t('player aid')"></i>
+    </div>
+  </a>
+
+  <div class="preferences_item preferences_item--settings">
+    <i class="preferences_icon preferences_icon--settings" :class="{'preferences_item--is-active': ui.preferences_panel_open}" v-on:click="ui.preferences_panel_open = !ui.preferences_panel_open"></i>
+    <preferences-dialog v-if="ui.preferences_panel_open" @okButtonClicked="ui.preferences_panel_open = false"/>
   </div>
 </div>
     `,
